chore: add wrapper for sentence collection layout
