--- web/src/components/pages/contribution/review/review.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react'
-
-const Review = () => <h1>Review page</h1>
-
-export default Review
