+++ web/src/components/pages/contribution/sentence-collector/review/review.css
@@ -1,0 +1,21 @@
+.review-wrapper {
+    & .review-page-instruction {
+        display: flex;
+        width: 55%;
+        margin: 2vh auto 32px;
+
+        svg {
+            margin: 0 8px;
+
+            @media (--xs-down) {
+                width: 16px;
+                height: 16px;
+            }
+        }
+
+        @media (--xs-down) {
+            font-size: var(--font-size-xs);
+            margin-bottom: 16px;
+        }
+    }
+}
