+++ web/src/components/pages/contribution/sentence-collector/sentence-collector-wrapper.css
@@ -1,0 +1,46 @@
+@import url('../../../media.css');
+
+.write-page,.review-page {
+    width: 100%;
+
+    & .review-wrapper, .write-wrapper {
+        background-image: linear-gradient(
+            to top,
+            var(--white) 20%,
+            var(--desert-storm)
+        );
+        overflow: hidden;
+
+        @media (--xl-up) {
+            height: 80%;
+        }
+
+        & .write,.review {
+            margin: 0 auto;
+            width: 100%;
+            height: 100%;
+            max-width: var(--wide-desktop-width);
+
+            @media (--lg-up) {
+                padding: 50px 20px 10px;
+            }
+
+            @media (--sm-down) {
+                padding-bottom: 30px;
+            }
+        }
+    }
+
+    & .edit-icon {
+        path {
+            fill: var(--blue);
+        }
+    }
+
+    & .review-icon {
+        path {
+            fill: var(--light-purple);
+        }
+    }
+
+}
