+++ web/src/components/pages/contribution/sentence-collector/write/write.test.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { renderWithProviders } from '../../../../../test/render-with-providers'
+import { renderWithProviders } from '../../../../../../test/render-with-providers'
 import { fireEvent, screen, waitFor } from '@testing-library/react'
 
 import Write from './write'
